Relocated validate_input method. Renamed version_checker. Added check for if subject code exists in supported slugs
